fix: filter

+++ src/commands/moderator/SayCommand.ts
@@ -13,6 +13,8 @@ export class SayCommand extends Command {
     public async messageRun(message: Message, args: Args): Promise<Message> {
         const content: string = await args.rest("string");
 
+        if (!message.member.roles.cache.has("1068139995103244289")) return;
+
         if (content == null) return;
 
         return await message.delete().then(async () => {
